fas-agent/pyproject.toml
[tool.poetry]
name = "fas-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.32"}
pydantic = "<2"
rag-chroma = {path = "packages/rag-chroma", develop = true}
unstructured = "^0.11.0"
markdown = "^3.5.1"


[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
